requirements.txt
streamlit
pandas
langchain
langchain-google-genai
faiss-cpu
python-dotenv
